Modified SPI receive callback
In order to be able to measure more than 65535 packets over DMA the code was modified. In the SPI receive callback was added part, where the function "HAL_SPI_Receive_DMA" is called again, but with empty buffer, so it is possible to store 2* 65535 pakcets.

Only problem is that the DMA is set all ower again, when all the data are received.

Also, DMA low layer API was generated